Fix jsonAdapter extension functions on generics, demo KClass alternative
The existing implementation of `jsonAdapter` extension functions on `companion object`s is currently broken if the target class has generics. This fixes that, albeit in an API breaking way by removing the `types` parameter and using reified generics instead. Not sure how to best go about this, as the previous implementation with generics didn't work anyway. We could leave the `types` parameter, but that leaves an awkward API hole where a user could define the generic on `jsonAdapter` with different ones than what are passed in the `types` parameter. Note that the IDE marks the generic specification in the function body as unnecessary, but this appears to be a bug since it doesn't compile without it and I plan to file separately.

The alternative implementation proposed along with this is using `KClass` as the receiver for the extension function rather than rely on a `companion object`, similar to how Kotlinx-serialization works for its adapters. The IDE does mark the receiver as unnecessary, but it is necessary for scoping (and works wonderfully in the IDE autocomplete). As this is a very different API, there's a few ways we could go about it if we want to pursue it. Namely, the implementation I show here is a deprecation approach where a full deprecated annotation + autofix. This is nice because users can run a "fix-all" on their code base. Other approaches are to remove the `companion object` support entirely, or not use the `KClass` support at all.

A final option is to just remove this API all together, and leave it to a 3rd party library to do instead (though it would probably never be able to be stable since it would have to "predict" the names of the generated JsonAdapter classes)

Proposing this up for discussion/review/demo/bug report :)