[FEATURE] New factory API to control the type `labelKey` serialization.
New API `autoSerializeLabel(boolean)` added to allow the factory can be configured with boolean flag to disable automatic serialization of the type label, and it's respective value. Here is how the configuration may look like in the factory builder.

```java
PolymorphicJsonAdapterFactory.of(Message.class, "type")
    .withSubtype(Success.class, "success")
    .withSubtype(MessageWithType.class, "customType")
    .autoSerializeLabel(false)) // ← New API!
```

This is done for use cases when all known subtypes has serializable type property in the class definition, then it generates double `type` properties when serialized to JSON.