Hide more symbols from the public API
Hiding implementations of JsonReader and JsonWriter is
particularly annoying as we have to either move it out
of the com.squareup.moshi package (and not have a sealed
class) or give it a mangled name.

I opted to give it a mangled name, though I'm not super
happy with this least-worst choice.